OpenPGP certificate needs either sign or encrypt capabilities
GnuTLS selects the OpenPGP certificate to use based on its
capabilities.  For ciphersuites doing key exchange via DHE (and
ECDHE), it wants the cert to be marked as usable for signing.  For
ciphersuites doing key exchange via RSA, it wants the key to be marked
as usable for encryption.

In our test suite, we'll mark the key as usable for signing.