Do not request X.509 certs when we are expecting OpenPGP
Apparently libgnutls26 on the server side does not deal properly with
the situation where it only has an OpenPGP certificate, and the client
sends an RFC 6091 extension that indicates (X509,OpenPGP) in that
order.  This may not be necessary if we link against libgnutls28.

It's OK to revert this if it works against libgnutls28, and we decide
to make mod_gnutls depend on libgnutls28 instead of libgnutls26.