clean up MGS_SIDE abuse of apr_pstrcat
We were allocating twice as many sections of RAM when exporting
environment variables as we needed to (the names of each env var as
well as the content).

This change statically allocates the names of most of the environment
variables, reducing the number of function calls and dynamic
allocations at runtime.