Let Automake handle the test directory
Instead of running the test script run_tests.sh in the top level
directory, move it to t/ and let make check call it there. Using
Automake in t/ requires moving the hand written makefile that manages
the actual tests out of the way. The result is that all test related
files are kept in t/, including test results.

This will make it easier to replace the monolithic test script
run_tests.sh with individual test cases managed by make.