Run monkeysphere-validation-agent only for the test case using it
Running and testing monkeysphere-validation-agent for each test case
causes tests to fail when it is not available, and unnecessary delays
when it is. Test cases using monkeysphere-validation-agent must set the
USE_MSVA environment variable to enable it.

Note that TestMakefile still calls the "pem2openpgp" tool from
Monkeysphere to set up PGP keys for the test environment.