runtests: Stop background sleep after gnutls-cli is done
The PID of the background sleep used to wait for an HTTP response is now
stored in a temporary file, so the sleep process can be stopped when it
is no longer needed. This avoids cluttering the system with sleep
processes and thus allows a more than generous timeout, which should be
useful for slow build systems.