Use server root for file paths in TLS proxy config
This is what users should be able to expect. Changing the return type of
load_proxy_x509_credentials to apr_status_t is not really a change: It
used APR status values before, but now the type is declared as such.
Absence of client credentials for the proxy is not necessarily a problem
(the back end server(s) may not require any), so the log level should be
lower than "warning".