configure: Try apr-util if apr_memcache library was not found
Fedora doesn't provide libapr_memcache, but apr-util contains the
necessary functions as well. With this commit ./configure checks for
apr_memcache first, and then for apr-util if it was not found.

This commit is inspired by a patch added to the Fedora package of
mod_gnutls by Jiri Kastner <jkastner@redhat.com> in version 0.5.10-1.
The original patch just replaces apr_memcache with apr-util.