Test suite: Prevent race over error log for server fail cases
In some test cases that expect the server to fail to start, it won't
even get far enough to write an error log. Sometimes these tests failed
because tail could not find the error log in the apache_down_err hook,
which shouldn't have been called at that point.

Two changes to the runtests script should remove the problem:

a) Call tail on the error log only if it is readable
b) Remove Apache error hook before trying to stop the server