Test suite: Start Apache directly again, search for httpd
Fedora installs a version of apachectl that does not pass arguments
through to Apache and relies on the system wide configuration. Thus
running the test suite there requires calling the Apache binary
directly. Search for it as "apache2" (name used in Debian) as well as
"httpd" (Fedora).

Also include test/cgi_module.conf and test/proxy_mods.conf in the
distribution archive.