Test suite: Wait loop for MSVA startup
Until now, there was a fixed wait time for MSVA startup. This caused one
of two problems: Either the wait time was short, risiking a test failure
due to MSVA not being ready in time, or it was long, making test runs
slow. Waiting just a short time but retrying for a longer time if
necessary helps.