Implement short message for mod_status reports
Until now the flags parameter from mod_status was ignored, and
mod_gnutls always returned an HTML report. With this change the flags
are checked for AP_STATUS_SHORT, and a non-HTML report is written if
it is set.

The formatting of the HTML report has been changed a little, too, so
the expected test output is updated accordingly.