configure: Check if creating namespaces is possible
Some Linux distributions (I've observed this on Debian) restrict
unprivileged users' ability to create namespaces, so checking just
whether the "unshare" command is available is not enough: If it is but
the user building mod_gnutls does not have the necessary permissions,
test cases will fail when trying to create their namespaces. Add a
check that tries to actually create a namespace and disable namespace
isolation of tests if it fails.