Register "ssl_is_https" function for compatibility with mod_rewrite
mod_rewrite calls this function to fill its %{HTTPS} special variable,
and not providing it meant that conditions like

  RewriteCond "%{HTTPS}" "off"

would match HTTPS connections using mod_gnutls. When used to redirect
clients from HTTP to HTTPS connections, this could lead to redirection
loops as reported in Debian bug #514005 [1]. In addition to
registering the function this commit also adds a test chase that
checks if an HTTP to HTTPS redirection works.

[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=514005