Test suite: Do not continue test case if Apache instance fails to start
On systems where namespaces aren't available, test cases in which
Apache HTTPD is expected not to start would sometimes fail when
running in parallel. The reason was a possible timing issue, where an
Apache instance for another test case might start before gnutls-cli is
run, and the TLS connection would unexpectedly succeed by connecting
to it.

Not attempting the TLS connection if HTTPD failed avoids this problem,
and also (slightly) speeds up tests.