Test suite: Do not explicitly set the mutex type to "default"
The setting can cause trouble when the mutex type "default" is file
based and its definition includes a path that the build process cannot
write to. This problem caused the Debian build to fail on hurd-i386,
where "default" resolved to "file:/var/run/apache2/" according to the
build log.

According to the HTTPD documentation a run-time file directory
relative to ServerRoot is used absent an explicit setting, and the
ServerRoot defined in test/base_apache.conf must be writable for the
test suite anyway.