Test suite: Set DefaultRuntimeDir for all Apache instances
According to the Apache documentation, the default run time directory
is relative to ServerRoot, which will be the build directory of the
mod_gnutls test suite. However, some Apache path layouts end up
setting DEFAULT_REL_RUNTIMEDIR to an absolute path, making this
statement wrong. If such a directory is not writable for the user
running "make check", tests may fail if they need the run time
directory, notably if using file-based mutexes as observed on the
Debian HURD build servers.