Remove broken SNI/session resumption workaround
By specification, the server name should be available to the post
client hello function, which uses it to select the virtual host. The
deleted "workaround" doesn't actually help if GnuTLS does not parse
the Server Name Indication extension correctly. As of this writing
there is a bug in GnuTLS that prevents SNI during resumption from
cache, while session tickets work (use "GnuTLSSessionTickets On").