mgs_get_ocsp_response(): Separate mgs_srvconf_rec* variable for readability
Having a separate variable instead of referencing ctxt every time
simplifies and shortens a lot of lines in this function. Technically
ctxt->sc is equal to the void *ptr parameter, but it's more explicit
this way, and also allows changing the custom pointer in the future
without changing the callback function.