Add proof-of-concept SNI parser in a pre client hello hook
The SNI parser is complete, but right now the hook only retrieves the
SNI data and logs it. The goal is to select the right virtual host and
load ALPN parameters (and possibly others) before GnuTLS processes the
ClientHello message. That should make different "Protocols" directives
between virtual hosts work as expected.