https-test-client.py: Implement minimal HTTP/1.0 client
The test case 33_vhost_SNI_serveralias_missinghost checks if the
server correctly rejects requests with SNI but no "Host"
header. However, an HTTP/1.1 request without a Host header is invalid
and the server code rejects it before mod_gnutls can check the
request. This means we must use HTTP/1.0 in that test to get a useful
result.