Test suite: Detect Apache shutdown by PID file
Wait for the PID file to disappear instead of using signal 0. The
reason is that some container environments don't have a proper init
system which reaps the Apache process on shutdown. In that case the
process stays around as a zombie process and os.kill() reports it as
existing indefinitely.