Use -VERS-ALL instead of -VERS-TLS-ALL
There was a bug in GnuTLS where leaving DTLS versions (which are
present in NORMAL) enabled could lead to a disabled TLS version being
accepted: https://gitlab.com/gnutls/gnutls/-/issues/1054

This has been fixed in the GnuTLS git repository (see
https://gitlab.com/gnutls/gnutls/-/merge_requests/1309), but there's
no release with the fix yet. I was testing with a local development
build so -VERS-TLS-ALL worked as it should, but the current distro
versions don't have the fix, so -VERS-ALL is needed.