Test successful and failed client auth with the same server instance
The remaining and removed server configurations were identical. The
difference was that in the "fail" the client did not authenticate (no
cert/key) and expected the connection to fail. This test has been
moved to the 10_basic_client_verification test as a second test
connection, achieving the same test coverage with one less round of
server start/stop.